created folder structure
