+++ services/backend/Makefile
@@ -1,0 +1,5 @@
+build:
+	go build -o bin/server cmd/server/main.go 
+
+run: build 
+	./bin/server
